refactor(ui): partially flatten `shared/components/` dirs
- whenever there was a redundant folder around a single component, flatten the directory structure
