+++ ui/src/app/cluster-workflow-templates/cluster-workflow-template-editor.tsx
@@ -5,7 +5,7 @@ import {WorkflowTemplate} from '../../models';
 import {LabelsAndAnnotationsEditor} from '../shared/components/editors/labels-and-annotations-editor';
 import {MetadataEditor} from '../shared/components/editors/metadata-editor';
 import {WorkflowParametersEditor} from '../shared/components/editors/workflow-parameters-editor';
-import {ObjectEditor} from '../shared/components/object-editor/object-editor';
+import {ObjectEditor} from '../shared/components/object-editor';
 
 export function ClusterWorkflowTemplateEditor({
     onChange,
