+++ ui/src/app/sensors/sensor-editor.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-
 import {Tabs} from 'argo-ui';
+
 import {Sensor} from '../../models';
 import {MetadataEditor} from '../shared/components/editors/metadata-editor';
-import {ObjectEditor} from '../shared/components/object-editor/object-editor';
+import {ObjectEditor} from '../shared/components/object-editor';
 
 export function SensorEditor({
     onChange,
