+++ ui/src/app/workflows/components/submit-workflow-panel.tsx
@@ -3,7 +3,7 @@ import React, {useMemo, useState} from 'react';
 import {Parameter, Template} from '../../../models';
 import {uiUrl} from '../../shared/base';
 import {ErrorNotice} from '../../shared/components/error-notice';
-import {ParametersInput} from '../../shared/components/parameters-input/parameters-input';
+import {ParametersInput} from '../../shared/components/parameters-input';
 import {TagsInput} from '../../shared/components/tags-input/tags-input';
 import {services} from '../../shared/services';
 import {Utils} from '../../shared/utils';
