+++ ui/src/app/workflows/components/workflow-editor.tsx
@@ -4,7 +4,7 @@ import {Tabs} from 'argo-ui';
 import {Workflow} from '../../../models';
 import {MetadataEditor} from '../../shared/components/editors/metadata-editor';
 import {WorkflowParametersEditor} from '../../shared/components/editors/workflow-parameters-editor';
-import {ObjectEditor} from '../../shared/components/object-editor/object-editor';
+import {ObjectEditor} from '../../shared/components/object-editor';
 
 export function WorkflowEditor({
     selectedTabKey,
